updated commonText and added commonMultiText
