+++ example/lib/screens/date_pickers.dart
@@ -20,14 +20,14 @@ class DatePickers extends StatelessWidget {
     return Scaffold(
       appBar: AppBar(
         backgroundColor: const Color(colorPrimary),
-        title: CommonText(
+        title: CommonMultiText(
           containerStyle: const CommonContainerModel().copyWith(
             backgroundColor: 0x00FFFFFF,
             alignment: Alignment.center,
             width: 0.7,
           ),
           style: CommonTextStyles().h2Style(),
-          text: 'DatePickers',
+          'DatePickers',
         ),
       ),
       body: CommonContainer(
@@ -65,7 +65,7 @@ class DatePickers extends StatelessWidget {
                   consoleLog(date);
                 },
               ),
-              CommonText(
+              CommonMultiText(
                 style: textStyle,
                 containerStyle: containerStyle,
                 onPress: () {
@@ -96,9 +96,9 @@ class DatePickers extends StatelessWidget {
                     currentTime: DateTime.now(),
                   );
                 },
-                text: 'show date picker(custom theme &date time range)',
+                'show date picker(custom theme &date time range)',
               ),
-              CommonText(
+              CommonMultiText(
                 style: textStyle,
                 containerStyle: containerStyle,
                 onPress: () {
@@ -110,9 +110,9 @@ class DatePickers extends StatelessWidget {
                     consoleLog('confirm $date');
                   }, currentTime: DateTime.now());
                 },
-                text: 'show time picker',
+                'show time picker',
               ),
-              CommonText(
+              CommonMultiText(
                 style: textStyle,
                 containerStyle: containerStyle,
                 onPress: () {
@@ -124,9 +124,9 @@ class DatePickers extends StatelessWidget {
                     consoleLog('confirm $date');
                   }, currentTime: DateTime.now());
                 },
-                text: 'show 12H time picker with AM/PM',
+                'show 12H time picker with AM/PM',
               ),
-              CommonText(
+              CommonMultiText(
                 style: textStyle,
                 containerStyle: containerStyle,
                 onPress: () {
@@ -140,9 +140,9 @@ class DatePickers extends StatelessWidget {
                     consoleLog('confirm $date');
                   }, locale: LocaleType.zh);
                 },
-                text: 'show date time picker (Chinese)',
+                'show date time picker (Chinese)',
               ),
-              CommonText(
+              CommonMultiText(
                 style: textStyle,
                 containerStyle: containerStyle,
                 onPress: () {
@@ -154,9 +154,9 @@ class DatePickers extends StatelessWidget {
                     consoleLog('confirm $date');
                   }, currentTime: DateTime(2008, 12, 31, 23, 12, 34));
                 },
-                text: 'show date time picker (English-America)',
+                'show date time picker (English-America)',
               ),
-              CommonText(
+              CommonMultiText(
                 style: textStyle,
                 containerStyle: containerStyle,
                 onPress: () {
@@ -170,9 +170,9 @@ class DatePickers extends StatelessWidget {
                       currentTime: DateTime(2008, 12, 31, 23, 12, 34),
                       locale: LocaleType.nl);
                 },
-                text: 'show date time picker (Dutch)',
+                'show date time picker (Dutch)',
               ),
-              CommonText(
+              CommonMultiText(
                 style: textStyle,
                 containerStyle: containerStyle,
                 onPress: () {
@@ -186,9 +186,9 @@ class DatePickers extends StatelessWidget {
                       currentTime: DateTime(2008, 12, 31, 23, 12, 34),
                       locale: LocaleType.ru);
                 },
-                text: 'show date time picker (Russian)',
+                'show date time picker (Russian)',
               ),
-              CommonText(
+              CommonMultiText(
                 style: textStyle,
                 containerStyle: containerStyle,
                 onPress: () {
@@ -202,9 +202,9 @@ class DatePickers extends StatelessWidget {
                       currentTime: DateTime.utc(2019, 12, 31, 23, 12, 34),
                       locale: LocaleType.de);
                 },
-                text: 'show date time picker in UTC (German)',
+                'show date time picker in UTC (German)',
               ),
-              CommonText(
+              CommonMultiText(
                 style: textStyle,
                 containerStyle: containerStyle,
                 onPress: () {
@@ -221,9 +221,9 @@ class DatePickers extends StatelessWidget {
                     locale: LocaleType.ar,
                   );
                 },
-                text: 'show date time picker in UTC (AR)',
+                'show date time picker in UTC (AR)',
               ),
-              CommonText(
+              CommonMultiText(
                 style: textStyle,
                 containerStyle: containerStyle,
                 onPress: () {
@@ -238,8 +238,7 @@ class DatePickers extends StatelessWidget {
                     },
                   );
                 },
-                text:
-                    'show custom time picker,\nyou can custom picker model like this',
+                'show custom time picker,\nyou can custom picker model like this',
               ),
             ],
           ),
