+++ example/lib/screens/index.dart
@@ -1,7 +1,6 @@
 export 'buttons.dart';
 export 'containers.dart';
 export 'date_pickers.dart';
-export 'drawer.dart';
 export 'images.dart';
 export 'text_inputs.dart';
 export 'texts.dart';
