+++ lib/src/components/common_date_picker_time_line/common_picker_timeline.dart
@@ -188,11 +188,11 @@ class _DatePickerState extends State<CommonPickerTimeLine> {
                   Icons.date_range_outlined,
                 ),
                 CommonText(
-                  text: '$currentMonth, ',
+                  '$currentMonth, ',
                   style: widget.headerTextStyle ?? defaultTextStyle,
                 ),
                 CommonText(
-                  text: currentYear,
+                  currentYear,
                   style: widget.headerTextStyle ?? defaultTextStyle,
                 ),
               ],
