+++ lib/src/components/common_multi_text/index.dart
@@ -1,0 +1,1 @@
+export 'common_multi_text.dart';
