--- lib/src/components/common_simple_text/index.dart
@@ -1,1 +1,0 @@
-export 'common_simple_text.dart';
