+++ lib/src/components/index.dart
@@ -2,8 +2,8 @@ export 'common_button/index.dart';
 export 'common_container/index.dart';
 export 'common_date_picker/index.dart';
 export 'common_date_picker_time_line/index.dart';
+export 'common_multi_text/index.dart';
 export 'common_old_button/index.dart';
-export 'common_simple_text/index.dart';
 export 'common_text/index.dart';
 export 'common_text_input/index.dart';
 export 'common_touchable/index.dart';
